requirements.txt
beautifulsoup4~=4.8.1
colorlog~=6.7.0
lxml~=4.9.1
PyYAML~=6.0.1
requests~=2.31.0
tqdm~=4.66.1
trakt.py~=4.4.0